move browserstack capabilities into cucumber config
