+++ playwright-staging/tests/support/hooks.js
@@ -1,35 +1,10 @@
 require('dotenv').config();
-
+const caps = require('../../config/browserstack');
 const { Before, After, Status, setDefaultTimeout } = require('@cucumber/cucumber');
 const { chromium } = require('@playwright/test');
 const Commands = require('../utils/commands');
 
 setDefaultTimeout(300 * 1000); // 5 mins
-
-// Get Playwright version to match BrowserStack runtime
-const clientPlaywrightVersion = require('@playwright/test/package.json').version;
-
-// BrowserStack capabilities
-const caps = {
-  project: 'giftaid-react',
-  name: 'e2e tests',
-  browser: 'chrome',
-  browser_version: 'latest',
-  resolution: '1920x1080',
-  os: 'Windows',
-  os_version: '11',
-  'browserstack.username': process.env.BROWSERSTACK_USERNAME,
-  'browserstack.accessKey': process.env.BROWSERSTACK_ACCESS_KEY,
-  // Ensure BrowserStack Playwright version matches local version
-  'client.playwrightVersion': clientPlaywrightVersion,
-  'browserstack.playwrightVersion': clientPlaywrightVersion,
-  
-  // logs
-  'browserstack.networkLogs': true,
-  'browserstack.console': 'info',
-  'browserstack.debug': true,
-  'browserstack.idleTimeout': 300,
-};
 
 // Runs before each scenario
 Before(async function (scenario) {
